+++ CHANGELOG.md
@@ -8,7 +8,7 @@ to a section with the version name.
 
 ## Unreleased Changes
 
-* Fix ZATCA Images Patch Fails by resize images.
+* Fix ZATCA desktop icon patch failure due to large size (> 1MiB). The image is now a much smaller SVG
 
 ## 0.60.0